Fix for trac ticket 1066 - Generic grid does is not positioned correctly within the grid boundary http://trac.osgeo.org/mapguide/ticket/1066
Submitted on behalf of Norm Olsen

Notes:
- Update from Norm due to Linux build failure.

git-svn-id: http://svn.osgeo.org/mapguide/trunk@4166 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Geometry/CoordinateSystem/CoordinateSystemFactory.cpp
@@ -650,7 +650,7 @@ MgCoordinateSystemGridBoundary* MgCoordinateSystemFactory::GridBoundary(MgPolygo
 MgCoordinateSystemMgrs* MgCoordinateSystemFactory::GetMgrs(
     double dEquatorialRadius, 
     double dEccentricity, 
-    int nLetteringScheme,
+    INT8 nLetteringScheme,
     bool bSetExceptionsOn)
 {
     Ptr<CCoordinateSystemMgrs> pNew;
